Moved export function to more sensible place
